@@ -1,604 +1,70 @@
-// Code generated by generator; DO NOT EDIT.
 package greenleaf
 
-import "time"
+const (
+	eqOp     = "$eq"
+	neOp     = "$ne"
+	gtOp     = "$gt"
+	gteOp    = "$gte"
+	ltOp     = "$lt"
+	lteOp    = "$lte"
+	inOp     = "$in"
+	existsOp = "$exists"
+	ninOp    = "$nin"
+)
 
-// EqBool ...
-func (f *FilterBuilder) EqBool(field string, value bool) *FilterBuilder {
-	return f.addSelector(field, "$eq", value)
+func Eq[T comparable](field string, value T) Document {
+	return createSelector(field, eqOp, value)
 }
 
-// EqFloat32 ...
-func (f *FilterBuilder) EqFloat32(field string, value float32) *FilterBuilder {
-	return f.addSelector(field, "$eq", value)
+func EqSlice[T any](field string, value []T) Document {
+	return createSelector(field, eqOp, value)
 }
 
-// EqFloat64 ...
-func (f *FilterBuilder) EqFloat64(field string, value float64) *FilterBuilder {
-	return f.addSelector(field, "$eq", value)
+func Ne[T comparable](field string, value T) Document {
+	return createSelector(field, neOp, value)
 }
 
-// EqInt ...
-func (f *FilterBuilder) EqInt(field string, value int) *FilterBuilder {
-	return f.addSelector(field, "$eq", value)
+func Gt[T comparable](field string, value T) Document {
+	return createSelector(field, gtOp, value)
 }
 
-// EqInt16 ...
-func (f *FilterBuilder) EqInt16(field string, value int16) *FilterBuilder {
-	return f.addSelector(field, "$eq", value)
+func Gte[T comparable](field string, value T) Document {
+	return createSelector(field, gteOp, value)
 }
 
-// EqInt32 ...
-func (f *FilterBuilder) EqInt32(field string, value int32) *FilterBuilder {
-	return f.addSelector(field, "$eq", value)
+func Lt[T comparable](field string, value T) Document {
+	return createSelector(field, ltOp, value)
 }
 
-// EqInt64 ...
-func (f *FilterBuilder) EqInt64(field string, value int64) *FilterBuilder {
-	return f.addSelector(field, "$eq", value)
+func Lte[T comparable](field string, value T) Document {
+	return createSelector(field, lteOp, value)
 }
 
-// EqInt8 ...
-func (f *FilterBuilder) EqInt8(field string, value int8) *FilterBuilder {
-	return f.addSelector(field, "$eq", value)
+func Exists(field string, value bool) Document {
+	return createSelector(field, existsOp, value)
 }
 
-// EqString ...
-func (f *FilterBuilder) EqString(field string, value string) *FilterBuilder {
-	return f.addSelector(field, "$eq", value)
+func In[T any](field string, value []T) Document {
+	return createSelector(field, inOp, value)
 }
 
-// EqUint ...
-func (f *FilterBuilder) EqUint(field string, value uint) *FilterBuilder {
-	return f.addSelector(field, "$eq", value)
+func Nin[T any](field string, value []T) Document {
+	return createSelector(field, ninOp, value)
 }
 
-// EqUint16 ...
-func (f *FilterBuilder) EqUint16(field string, value uint16) *FilterBuilder {
-	return f.addSelector(field, "$eq", value)
+func createSelector(field, operator string, value any) Document {
+	return Document{
+		field: M{operator: value},
+	}
 }
 
-// EqUint32 ...
-func (f *FilterBuilder) EqUint32(field string, value uint32) *FilterBuilder {
-	return f.addSelector(field, "$eq", value)
-}
-
-// EqUint64 ...
-func (f *FilterBuilder) EqUint64(field string, value uint64) *FilterBuilder {
-	return f.addSelector(field, "$eq", value)
-}
-
-// EqUint8 ...
-func (f *FilterBuilder) EqUint8(field string, value uint8) *FilterBuilder {
-	return f.addSelector(field, "$eq", value)
-}
-
-// EqTime ...
-func (f *FilterBuilder) EqTime(field string, value time.Time) *FilterBuilder {
-	return f.addSelector(field, "$eq", value)
-}
-
-// NeBool ...
-func (f *FilterBuilder) NeBool(field string, value bool) *FilterBuilder {
-	return f.addSelector(field, "$ne", value)
-}
-
-// NeFloat32 ...
-func (f *FilterBuilder) NeFloat32(field string, value float32) *FilterBuilder {
-	return f.addSelector(field, "$ne", value)
-}
-
-// NeFloat64 ...
-func (f *FilterBuilder) NeFloat64(field string, value float64) *FilterBuilder {
-	return f.addSelector(field, "$ne", value)
-}
-
-// NeInt ...
-func (f *FilterBuilder) NeInt(field string, value int) *FilterBuilder {
-	return f.addSelector(field, "$ne", value)
-}
-
-// NeInt16 ...
-func (f *FilterBuilder) NeInt16(field string, value int16) *FilterBuilder {
-	return f.addSelector(field, "$ne", value)
-}
-
-// NeInt32 ...
-func (f *FilterBuilder) NeInt32(field string, value int32) *FilterBuilder {
-	return f.addSelector(field, "$ne", value)
-}
-
-// NeInt64 ...
-func (f *FilterBuilder) NeInt64(field string, value int64) *FilterBuilder {
-	return f.addSelector(field, "$ne", value)
-}
-
-// NeInt8 ...
-func (f *FilterBuilder) NeInt8(field string, value int8) *FilterBuilder {
-	return f.addSelector(field, "$ne", value)
-}
-
-// NeString ...
-func (f *FilterBuilder) NeString(field string, value string) *FilterBuilder {
-	return f.addSelector(field, "$ne", value)
-}
-
-// NeUint ...
-func (f *FilterBuilder) NeUint(field string, value uint) *FilterBuilder {
-	return f.addSelector(field, "$ne", value)
-}
-
-// NeUint16 ...
-func (f *FilterBuilder) NeUint16(field string, value uint16) *FilterBuilder {
-	return f.addSelector(field, "$ne", value)
-}
-
-// NeUint32 ...
-func (f *FilterBuilder) NeUint32(field string, value uint32) *FilterBuilder {
-	return f.addSelector(field, "$ne", value)
-}
-
-// NeUint64 ...
-func (f *FilterBuilder) NeUint64(field string, value uint64) *FilterBuilder {
-	return f.addSelector(field, "$ne", value)
-}
-
-// NeUint8 ...
-func (f *FilterBuilder) NeUint8(field string, value uint8) *FilterBuilder {
-	return f.addSelector(field, "$ne", value)
-}
-
-// NeTime ...
-func (f *FilterBuilder) NeTime(field string, value time.Time) *FilterBuilder {
-	return f.addSelector(field, "$ne", value)
-}
-
-// GtBool ...
-func (f *FilterBuilder) GtBool(field string, value bool) *FilterBuilder {
-	return f.addSelector(field, "$gt", value)
-}
-
-// GtFloat32 ...
-func (f *FilterBuilder) GtFloat32(field string, value float32) *FilterBuilder {
-	return f.addSelector(field, "$gt", value)
-}
-
-// GtFloat64 ...
-func (f *FilterBuilder) GtFloat64(field string, value float64) *FilterBuilder {
-	return f.addSelector(field, "$gt", value)
-}
-
-// GtInt ...
-func (f *FilterBuilder) GtInt(field string, value int) *FilterBuilder {
-	return f.addSelector(field, "$gt", value)
-}
-
-// GtInt16 ...
-func (f *FilterBuilder) GtInt16(field string, value int16) *FilterBuilder {
-	return f.addSelector(field, "$gt", value)
-}
-
-// GtInt32 ...
-func (f *FilterBuilder) GtInt32(field string, value int32) *FilterBuilder {
-	return f.addSelector(field, "$gt", value)
-}
-
-// GtInt64 ...
-func (f *FilterBuilder) GtInt64(field string, value int64) *FilterBuilder {
-	return f.addSelector(field, "$gt", value)
-}
-
-// GtInt8 ...
-func (f *FilterBuilder) GtInt8(field string, value int8) *FilterBuilder {
-	return f.addSelector(field, "$gt", value)
-}
-
-// GtString ...
-func (f *FilterBuilder) GtString(field string, value string) *FilterBuilder {
-	return f.addSelector(field, "$gt", value)
-}
-
-// GtUint ...
-func (f *FilterBuilder) GtUint(field string, value uint) *FilterBuilder {
-	return f.addSelector(field, "$gt", value)
-}
-
-// GtUint16 ...
-func (f *FilterBuilder) GtUint16(field string, value uint16) *FilterBuilder {
-	return f.addSelector(field, "$gt", value)
-}
-
-// GtUint32 ...
-func (f *FilterBuilder) GtUint32(field string, value uint32) *FilterBuilder {
-	return f.addSelector(field, "$gt", value)
-}
-
-// GtUint64 ...
-func (f *FilterBuilder) GtUint64(field string, value uint64) *FilterBuilder {
-	return f.addSelector(field, "$gt", value)
-}
-
-// GtUint8 ...
-func (f *FilterBuilder) GtUint8(field string, value uint8) *FilterBuilder {
-	return f.addSelector(field, "$gt", value)
-}
-
-// GtTime ...
-func (f *FilterBuilder) GtTime(field string, value time.Time) *FilterBuilder {
-	return f.addSelector(field, "$gt", value)
-}
-
-// GteBool ...
-func (f *FilterBuilder) GteBool(field string, value bool) *FilterBuilder {
-	return f.addSelector(field, "$gte", value)
-}
-
-// GteFloat32 ...
-func (f *FilterBuilder) GteFloat32(field string, value float32) *FilterBuilder {
-	return f.addSelector(field, "$gte", value)
-}
-
-// GteFloat64 ...
-func (f *FilterBuilder) GteFloat64(field string, value float64) *FilterBuilder {
-	return f.addSelector(field, "$gte", value)
-}
-
-// GteInt ...
-func (f *FilterBuilder) GteInt(field string, value int) *FilterBuilder {
-	return f.addSelector(field, "$gte", value)
-}
-
-// GteInt16 ...
-func (f *FilterBuilder) GteInt16(field string, value int16) *FilterBuilder {
-	return f.addSelector(field, "$gte", value)
-}
-
-// GteInt32 ...
-func (f *FilterBuilder) GteInt32(field string, value int32) *FilterBuilder {
-	return f.addSelector(field, "$gte", value)
-}
-
-// GteInt64 ...
-func (f *FilterBuilder) GteInt64(field string, value int64) *FilterBuilder {
-	return f.addSelector(field, "$gte", value)
-}
-
-// GteInt8 ...
-func (f *FilterBuilder) GteInt8(field string, value int8) *FilterBuilder {
-	return f.addSelector(field, "$gte", value)
-}
-
-// GteString ...
-func (f *FilterBuilder) GteString(field string, value string) *FilterBuilder {
-	return f.addSelector(field, "$gte", value)
-}
-
-// GteUint ...
-func (f *FilterBuilder) GteUint(field string, value uint) *FilterBuilder {
-	return f.addSelector(field, "$gte", value)
-}
-
-// GteUint16 ...
-func (f *FilterBuilder) GteUint16(field string, value uint16) *FilterBuilder {
-	return f.addSelector(field, "$gte", value)
-}
-
-// GteUint32 ...
-func (f *FilterBuilder) GteUint32(field string, value uint32) *FilterBuilder {
-	return f.addSelector(field, "$gte", value)
-}
-
-// GteUint64 ...
-func (f *FilterBuilder) GteUint64(field string, value uint64) *FilterBuilder {
-	return f.addSelector(field, "$gte", value)
-}
-
-// GteUint8 ...
-func (f *FilterBuilder) GteUint8(field string, value uint8) *FilterBuilder {
-	return f.addSelector(field, "$gte", value)
-}
-
-// GteTime ...
-func (f *FilterBuilder) GteTime(field string, value time.Time) *FilterBuilder {
-	return f.addSelector(field, "$gte", value)
-}
-
-// LtBool ...
-func (f *FilterBuilder) LtBool(field string, value bool) *FilterBuilder {
-	return f.addSelector(field, "$lt", value)
-}
-
-// LtFloat32 ...
-func (f *FilterBuilder) LtFloat32(field string, value float32) *FilterBuilder {
-	return f.addSelector(field, "$lt", value)
-}
-
-// LtFloat64 ...
-func (f *FilterBuilder) LtFloat64(field string, value float64) *FilterBuilder {
-	return f.addSelector(field, "$lt", value)
-}
-
-// LtInt ...
-func (f *FilterBuilder) LtInt(field string, value int) *FilterBuilder {
-	return f.addSelector(field, "$lt", value)
-}
-
-// LtInt16 ...
-func (f *FilterBuilder) LtInt16(field string, value int16) *FilterBuilder {
-	return f.addSelector(field, "$lt", value)
-}
-
-// LtInt32 ...
-func (f *FilterBuilder) LtInt32(field string, value int32) *FilterBuilder {
-	return f.addSelector(field, "$lt", value)
-}
-
-// LtInt64 ...
-func (f *FilterBuilder) LtInt64(field string, value int64) *FilterBuilder {
-	return f.addSelector(field, "$lt", value)
-}
-
-// LtInt8 ...
-func (f *FilterBuilder) LtInt8(field string, value int8) *FilterBuilder {
-	return f.addSelector(field, "$lt", value)
-}
-
-// LtString ...
-func (f *FilterBuilder) LtString(field string, value string) *FilterBuilder {
-	return f.addSelector(field, "$lt", value)
-}
-
-// LtUint ...
-func (f *FilterBuilder) LtUint(field string, value uint) *FilterBuilder {
-	return f.addSelector(field, "$lt", value)
-}
-
-// LtUint16 ...
-func (f *FilterBuilder) LtUint16(field string, value uint16) *FilterBuilder {
-	return f.addSelector(field, "$lt", value)
-}
-
-// LtUint32 ...
-func (f *FilterBuilder) LtUint32(field string, value uint32) *FilterBuilder {
-	return f.addSelector(field, "$lt", value)
-}
-
-// LtUint64 ...
-func (f *FilterBuilder) LtUint64(field string, value uint64) *FilterBuilder {
-	return f.addSelector(field, "$lt", value)
-}
-
-// LtUint8 ...
-func (f *FilterBuilder) LtUint8(field string, value uint8) *FilterBuilder {
-	return f.addSelector(field, "$lt", value)
-}
-
-// LtTime ...
-func (f *FilterBuilder) LtTime(field string, value time.Time) *FilterBuilder {
-	return f.addSelector(field, "$lt", value)
-}
-
-// LteBool ...
-func (f *FilterBuilder) LteBool(field string, value bool) *FilterBuilder {
-	return f.addSelector(field, "$lte", value)
-}
-
-// LteFloat32 ...
-func (f *FilterBuilder) LteFloat32(field string, value float32) *FilterBuilder {
-	return f.addSelector(field, "$lte", value)
-}
-
-// LteFloat64 ...
-func (f *FilterBuilder) LteFloat64(field string, value float64) *FilterBuilder {
-	return f.addSelector(field, "$lte", value)
-}
-
-// LteInt ...
-func (f *FilterBuilder) LteInt(field string, value int) *FilterBuilder {
-	return f.addSelector(field, "$lte", value)
-}
-
-// LteInt16 ...
-func (f *FilterBuilder) LteInt16(field string, value int16) *FilterBuilder {
-	return f.addSelector(field, "$lte", value)
-}
-
-// LteInt32 ...
-func (f *FilterBuilder) LteInt32(field string, value int32) *FilterBuilder {
-	return f.addSelector(field, "$lte", value)
-}
-
-// LteInt64 ...
-func (f *FilterBuilder) LteInt64(field string, value int64) *FilterBuilder {
-	return f.addSelector(field, "$lte", value)
-}
-
-// LteInt8 ...
-func (f *FilterBuilder) LteInt8(field string, value int8) *FilterBuilder {
-	return f.addSelector(field, "$lte", value)
-}
-
-// LteString ...
-func (f *FilterBuilder) LteString(field string, value string) *FilterBuilder {
-	return f.addSelector(field, "$lte", value)
-}
-
-// LteUint ...
-func (f *FilterBuilder) LteUint(field string, value uint) *FilterBuilder {
-	return f.addSelector(field, "$lte", value)
-}
-
-// LteUint16 ...
-func (f *FilterBuilder) LteUint16(field string, value uint16) *FilterBuilder {
-	return f.addSelector(field, "$lte", value)
-}
-
-// LteUint32 ...
-func (f *FilterBuilder) LteUint32(field string, value uint32) *FilterBuilder {
-	return f.addSelector(field, "$lte", value)
-}
-
-// LteUint64 ...
-func (f *FilterBuilder) LteUint64(field string, value uint64) *FilterBuilder {
-	return f.addSelector(field, "$lte", value)
-}
-
-// LteUint8 ...
-func (f *FilterBuilder) LteUint8(field string, value uint8) *FilterBuilder {
-	return f.addSelector(field, "$lte", value)
-}
-
-// LteTime ...
-func (f *FilterBuilder) LteTime(field string, value time.Time) *FilterBuilder {
-	return f.addSelector(field, "$lte", value)
-}
-
-// InBool ...
-func (f *FilterBuilder) InBool(field string, value []bool) *FilterBuilder {
-	return f.addSelector(field, "$in", value)
-}
-
-// InFloat32 ...
-func (f *FilterBuilder) InFloat32(field string, value []float32) *FilterBuilder {
-	return f.addSelector(field, "$in", value)
-}
-
-// InFloat64 ...
-func (f *FilterBuilder) InFloat64(field string, value []float64) *FilterBuilder {
-	return f.addSelector(field, "$in", value)
-}
-
-// InInt ...
-func (f *FilterBuilder) InInt(field string, value []int) *FilterBuilder {
-	return f.addSelector(field, "$in", value)
-}
-
-// InInt16 ...
-func (f *FilterBuilder) InInt16(field string, value []int16) *FilterBuilder {
-	return f.addSelector(field, "$in", value)
-}
-
-// InInt32 ...
-func (f *FilterBuilder) InInt32(field string, value []int32) *FilterBuilder {
-	return f.addSelector(field, "$in", value)
-}
-
-// InInt64 ...
-func (f *FilterBuilder) InInt64(field string, value []int64) *FilterBuilder {
-	return f.addSelector(field, "$in", value)
-}
-
-// InInt8 ...
-func (f *FilterBuilder) InInt8(field string, value []int8) *FilterBuilder {
-	return f.addSelector(field, "$in", value)
-}
-
-// InString ...
-func (f *FilterBuilder) InString(field string, value []string) *FilterBuilder {
-	return f.addSelector(field, "$in", value)
-}
-
-// InUint ...
-func (f *FilterBuilder) InUint(field string, value []uint) *FilterBuilder {
-	return f.addSelector(field, "$in", value)
-}
-
-// InUint16 ...
-func (f *FilterBuilder) InUint16(field string, value []uint16) *FilterBuilder {
-	return f.addSelector(field, "$in", value)
-}
-
-// InUint32 ...
-func (f *FilterBuilder) InUint32(field string, value []uint32) *FilterBuilder {
-	return f.addSelector(field, "$in", value)
-}
-
-// InUint64 ...
-func (f *FilterBuilder) InUint64(field string, value []uint64) *FilterBuilder {
-	return f.addSelector(field, "$in", value)
-}
-
-// InUint8 ...
-func (f *FilterBuilder) InUint8(field string, value []uint8) *FilterBuilder {
-	return f.addSelector(field, "$in", value)
-}
-
-// InTime ...
-func (f *FilterBuilder) InTime(field string, value time.Time) *FilterBuilder {
-	return f.addSelector(field, "$in", value)
-}
-
-// NinBool ...
-func (f *FilterBuilder) NinBool(field string, value []bool) *FilterBuilder {
-	return f.addSelector(field, "$nin", value)
-}
-
-// NinFloat32 ...
-func (f *FilterBuilder) NinFloat32(field string, value []float32) *FilterBuilder {
-	return f.addSelector(field, "$nin", value)
-}
-
-// NinFloat64 ...
-func (f *FilterBuilder) NinFloat64(field string, value []float64) *FilterBuilder {
-	return f.addSelector(field, "$nin", value)
-}
-
-// NinInt ...
-func (f *FilterBuilder) NinInt(field string, value []int) *FilterBuilder {
-	return f.addSelector(field, "$nin", value)
-}
-
-// NinInt16 ...
-func (f *FilterBuilder) NinInt16(field string, value []int16) *FilterBuilder {
-	return f.addSelector(field, "$nin", value)
-}
-
-// NinInt32 ...
-func (f *FilterBuilder) NinInt32(field string, value []int32) *FilterBuilder {
-	return f.addSelector(field, "$nin", value)
-}
-
-// NinInt64 ...
-func (f *FilterBuilder) NinInt64(field string, value []int64) *FilterBuilder {
-	return f.addSelector(field, "$nin", value)
-}
-
-// NinInt8 ...
-func (f *FilterBuilder) NinInt8(field string, value []int8) *FilterBuilder {
-	return f.addSelector(field, "$nin", value)
-}
-
-// NinString ...
-func (f *FilterBuilder) NinString(field string, value []string) *FilterBuilder {
-	return f.addSelector(field, "$nin", value)
-}
-
-// NinUint ...
-func (f *FilterBuilder) NinUint(field string, value []uint) *FilterBuilder {
-	return f.addSelector(field, "$nin", value)
-}
-
-// NinUint16 ...
-func (f *FilterBuilder) NinUint16(field string, value []uint16) *FilterBuilder {
-	return f.addSelector(field, "$nin", value)
-}
-
-// NinUint32 ...
-func (f *FilterBuilder) NinUint32(field string, value []uint32) *FilterBuilder {
-	return f.addSelector(field, "$nin", value)
-}
-
-// NinUint64 ...
-func (f *FilterBuilder) NinUint64(field string, value []uint64) *FilterBuilder {
-	return f.addSelector(field, "$nin", value)
-}
-
-// NinUint8 ...
-func (f *FilterBuilder) NinUint8(field string, value []uint8) *FilterBuilder {
-	return f.addSelector(field, "$nin", value)
-}
+func Filter(docs ...Document) Document {
+	query := Document{}
+	for _, doc := range docs {
+		for k, v := range doc {
+			query[k] = v
+		}
+	}
 
-// NinTime ...
-func (f *FilterBuilder) NinTime(field string, value time.Time) *FilterBuilder {
-	return f.addSelector(field, "$nin", value)
+	return query
 }
